Refactor ReadConfigurationBuilder
