+++ janusgraph-core/src/main/java/org/janusgraph/diskstorage/Backend.java
@@ -25,6 +25,7 @@ import org.janusgraph.core.JanusGraphException;
 import org.janusgraph.core.schema.JanusGraphManagement;
 import org.janusgraph.diskstorage.configuration.*;
 import org.janusgraph.diskstorage.configuration.backend.CommonsConfiguration;
+import org.janusgraph.diskstorage.configuration.backend.builder.KCVSConfigurationBuilder;
 import org.janusgraph.diskstorage.idmanagement.ConsistentKeyIDAuthority;
 import org.janusgraph.diskstorage.indexing.*;
 import org.janusgraph.diskstorage.keycolumnvalue.*;
@@ -47,7 +48,6 @@ import org.janusgraph.diskstorage.configuration.backend.KCVSConfiguration;
 import org.janusgraph.diskstorage.util.MetricInstrumentedStoreManager;
 import org.janusgraph.diskstorage.util.StandardBaseTransactionConfig;
 import org.janusgraph.diskstorage.util.time.TimestampProvider;
-import org.janusgraph.graphdb.configuration.GraphDatabaseConfiguration;
 import org.janusgraph.graphdb.transaction.TransactionConfiguration;
 import org.janusgraph.util.system.ConfigurationUtil;
 
@@ -275,7 +275,8 @@ public class Backend implements LockerProvider, AutoCloseable {
 
             //Open global configuration
             KeyColumnValueStore systemConfigStore = storeManagerLocking.openDatabase(SYSTEM_PROPERTIES_STORE_NAME);
-            systemConfig = getGlobalConfiguration(new BackendOperation.TransactionalProvider() {
+            KCVSConfigurationBuilder kcvsConfigurationBuilder = new KCVSConfigurationBuilder();
+            systemConfig = kcvsConfigurationBuilder.buildGlobalConfiguration(new BackendOperation.TransactionalProvider() {
                 @Override
                 public StoreTransaction openTx() throws BackendException {
                     return storeManagerLocking.beginTransaction(StandardBaseTransactionConfig.of(
@@ -288,7 +289,7 @@ public class Backend implements LockerProvider, AutoCloseable {
                     //Do nothing, storeManager is closed explicitly by Backend
                 }
             },systemConfigStore,configuration);
-            userConfig = getConfiguration(new BackendOperation.TransactionalProvider() {
+            userConfig = kcvsConfigurationBuilder.buildConfiguration(new BackendOperation.TransactionalProvider() {
                 @Override
                 public StoreTransaction openTx() throws BackendException {
                     return storeManagerLocking.beginTransaction(StandardBaseTransactionConfig.of(configuration.get(TIMESTAMP_PROVIDER)));
@@ -416,45 +417,6 @@ public class Backend implements LockerProvider, AutoCloseable {
         }
         Preconditions.checkArgument(manager instanceof KeyColumnValueStoreManager,"Invalid storage manager: %s",manager.getClass());
         return (KeyColumnValueStoreManager) manager;
-    }
-
-    private static KCVSConfiguration getGlobalConfiguration(final BackendOperation.TransactionalProvider txProvider,
-                                                                     final KeyColumnValueStore store,
-                                                                     final Configuration config) {
-        return getConfiguration(txProvider, store, SYSTEM_CONFIGURATION_IDENTIFIER, config);
-    }
-
-    private static KCVSConfiguration getConfiguration(final BackendOperation.TransactionalProvider txProvider,
-                                                            final KeyColumnValueStore store, final String identifier,
-                                                            final Configuration config) {
-        try {
-            KCVSConfiguration keyColumnValueStoreConfiguration =
-                    new KCVSConfiguration(txProvider,config,store,identifier);
-            keyColumnValueStoreConfiguration.setMaxOperationWaitTime(config.get(SETUP_WAITTIME));
-            return keyColumnValueStoreConfiguration;
-        } catch (BackendException e) {
-            throw new JanusGraphException("Could not open global configuration",e);
-        }
-    }
-
-    public static KCVSConfiguration getStandaloneGlobalConfiguration(final KeyColumnValueStoreManager manager,
-                                                                     final Configuration config) {
-        try {
-            final StoreFeatures features = manager.getFeatures();
-            return getGlobalConfiguration(new BackendOperation.TransactionalProvider() {
-                @Override
-                public StoreTransaction openTx() throws BackendException {
-                    return manager.beginTransaction(StandardBaseTransactionConfig.of(config.get(TIMESTAMP_PROVIDER),features.getKeyConsistentTxConfig()));
-                }
-
-                @Override
-                public void close() throws BackendException {
-                    manager.close();
-                }
-            },manager.openDatabase(SYSTEM_PROPERTIES_STORE_NAME),config);
-        } catch (BackendException e) {
-            throw new JanusGraphException("Could not open global configuration",e);
-        }
     }
 
     private static Map<String, IndexProvider> getIndexes(Configuration config) {
